services/db/human/requirements.txt
fastapi==0.116.0
psycopg2-binary==2.9.10
pydantic==2.11.7
sqlalchemy==2.0.41
uvicorn==0.29.0 
python-dotenv==1.1.1
alembic==1.16.4
pydantic_core==2.33.2
pytz==2025.2
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
email_validator==2.2.0